+++ pkg/upload/all.go
@@ -4,7 +4,6 @@ import (
 	"github.com/bugsnag/bugsnag-cli/pkg/log"
 	"github.com/bugsnag/bugsnag-cli/pkg/server"
 	"github.com/bugsnag/bugsnag-cli/pkg/utils"
-	"path/filepath"
 )
 
 type DiscoverAndUploadAny struct {
@@ -12,14 +11,21 @@ type DiscoverAndUploadAny struct {
 	UploadOptions map[string]string `help:"Additional arguments to pass to the upload request" mapsep:","`
 }
 
-func All(paths []string, options map[string]string, endpoint string, timeout int, retries int, overwrite bool,
-	apiKey string, failOnUploadError bool, dryRun bool) error {
+func All(
+	paths []string,
+	options map[string]string,
+	endpoint string,
+	timeout int,
+	retries int,
+	overwrite bool,
+	apiKey string,
+	dryRun bool,
+) error {
 
 	// Build the file list from the path(s)
 	log.Info("building file list...")
 
 	fileList, err := utils.BuildFileList(paths)
-	numberOfFiles := len(fileList)
 
 	if err != nil {
 		log.Error(" error building file list", 1)
@@ -51,20 +57,11 @@ func All(paths []string, options map[string]string, endpoint string, timeout int
 			fileFieldData["file"] = file
 		}
 
-		requestStatus := server.ProcessFileRequest(endpoint, uploadOptions, fileFieldData, timeout, retries, file, dryRun)
+		err := server.ProcessFileRequest(endpoint, uploadOptions, fileFieldData, timeout, retries, file, dryRun)
 
-		if requestStatus != nil {
-			if numberOfFiles > 1 {
-				if failOnUploadError {
-					return err
-				} else {
-					log.Warn(err.Error())
-				}
-			} else {
-				return err
-			}
-		} else {
-			log.Success("Uploaded " + filepath.Base(file))
+		if err != nil {
+
+			return err
 		}
 	}